+++ packages/@uppy/aws-s3/src/index.test.ts
@@ -29,7 +29,7 @@ describe('AwsS3Multipart', () => {
       core.use(AwsS3Multipart)
       const awsS3Multipart = core.getPlugin('AwsS3Multipart')!
 
-      const err = 'Expected a `companionUrl` option'
+      const err = 'Expected a `endpoint` option'
       const file = {}
       const opts = {}
 
@@ -91,6 +91,7 @@ describe('AwsS3Multipart', () => {
       expect(uploadSuccessHandler.mock.calls[0][1]).toStrictEqual({
         body: {
           ETag: 'test',
+          etag: 'test',
           location: 'http://example.com',
         },
         status: 200,
@@ -173,7 +174,9 @@ describe('AwsS3Multipart', () => {
         createMultipartUpload: vi.fn((file) => {
           // @ts-expect-error protected property
           const multipartUploader = awsS3Multipart.uploaders[file.id]!
-          const testChunkState = multipartUploader.chunkState[6]
+          const testChunkState =
+            // @ts-expect-error private method
+            multipartUploader[Symbol.for('uppy test: getChunkState')]()[6]
           let busy = false
           let done = false
           busySpy = vi.fn((value) => {
@@ -327,8 +330,8 @@ describe('AwsS3Multipart', () => {
     beforeEach(() => {
       core = new Core<any, Body>()
       core.use(AwsS3Multipart, {
-        companionUrl: '',
-        companionHeaders: {
+        endpoint: '',
+        headers: {
           authorization: oldToken,
         },
       })
@@ -337,7 +340,8 @@ describe('AwsS3Multipart', () => {
 
     it('companionHeader is updated before uploading file', async () => {
       awsS3Multipart.setOptions({
-        companionHeaders: {
+        endpoint: 'http://localhost',
+        headers: {
           authorization: newToken,
         },
       })
@@ -368,7 +372,8 @@ describe('AwsS3Multipart', () => {
           Body
         >
         awsS3Multipart.setOptions({
-          companionHeaders: {
+          endpoint: 'http://localhost',
+          headers: {
             authorization: newToken,
           },
         })