+++ packages/@uppy/google-drive/src/GoogleDrive.tsx
@@ -34,6 +34,8 @@ export default class GoogleDrive<
 
   files: UppyFile<M, B>[]
 
+  rootFolderId: string | null = 'root'
+
   constructor(uppy: Uppy<M, B>, opts: GoogleDriveOptions) {
     super(uppy, opts)
     this.type = 'acquirer'
@@ -96,7 +98,6 @@ export default class GoogleDrive<
     this.i18nInit()
     this.title = this.i18n('pluginNameGoogleDrive')
 
-    this.onFirstRender = this.onFirstRender.bind(this)
     this.render = this.render.bind(this)
   }
 
@@ -115,13 +116,6 @@ export default class GoogleDrive<
   uninstall(): void {
     this.view.tearDown()
     this.unmount()
-  }
-
-  async onFirstRender(): Promise<void> {
-    await Promise.all([
-      this.provider.fetchPreAuthToken(),
-      this.view.getFolder('root'),
-    ])
   }
 
   render(state: unknown): ComponentChild {