+++ packages/@uppy/utils/src/FileProgress.ts
@@ -28,7 +28,6 @@ export type FileProgressStarted = FileProgressBase & {
 }
 export type FileProgressNotStarted = FileProgressBase & {
   uploadStarted: null
-  // TODO: remove `|0` (or maybe `false|`?)
-  bytesUploaded: false | 0
+  bytesUploaded: false
 }
 export type FileProgress = FileProgressStarted | FileProgressNotStarted