+++ packages/@uppy/aws-s3/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @uppy/aws-s3
 
+## 4.0.0-beta.5
+
+Released: 2024-06-04
+Included in: Uppy v4.0.0-beta.10
+
+- @uppy/aws-s3: remove todo (Mikael Finstad / #5200)
+- @uppy/aws-s3: do not expose internal `assertHost` method (Mikael Finstad / #5200)
+- @uppy/aws-s3: make passing `signal` consistent (Mikael Finstad / #5200)
+- @uppy/aws-s3: remove `chunkState` getter (Mikael Finstad / #5200)
+- @uppy/aws-s3: remove `uploaderSockets` (Mikael Finstad / #5200)
+- @uppy/aws-s3,@uppy/tus,@uppy/utils,@uppy/xhr-upload: remove `uploader` from `upload-progress` event (Mikael Finstad / #5200)
+
 ## 4.0.0-beta.2
 
 Released: 2024-04-11